renamed nametuple
